requirements.txt
defusedxml==0.6.0
Django==2.1.1
djangorestframework==3.11.0
djangorestframework-xml==1.4.0
mysqlclient==1.4.6
pytz==2019.3
